Add other brands, fix switches, restructure code
* Other brands than (Purmo) CleverTouch are now supported. They
have not been tested with any real homes/devices so use with
caution and expect minor glitches, like missing or extra
features avialable compared to the actual device.

* Fix switches - there doesn't seem to exist a completely reliable way
to tell whether a device supporting on/off is a light or an electrical
outlet (or something else). More devices types are therefore now simply
represented as switches.

* While implementing support for the above, code was restructured where
it was deemed appropriate. This unfortunately risk introducing new
bugs, so the integration is obviously still in beta (or alpha).

* Supporting new providers while maintaining old unique ids was not
worth the effort. Users might need to remove and readd the integration.